RAG/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo[srv]==4.5.0
openai==0.27.8
langchain==0.0.304
numpy==1.24.3
gensim==4.3.0
python-docx==0.8.11
pillow==10.0.0
pymupdf==1.22.5
